refactor: Move main to top of project
