+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/sijis/glock/cmd/glock"
+
+func main() {
+	glock.Run()
+}
